[Refactor]Platform pin mem optimization
